issues/29 restructure rabbitmq publishing for better dependencies
