+++ src/Focus.Service.NewDayEmitter/Focus.Service.NewDayEmitter.csproj
@@ -16,5 +16,6 @@
   <ItemGroup>
     <ProjectReference Include="..\Focus.Core.Common\Focus.Core.Common.csproj" />
     <ProjectReference Include="..\Focus.Infrastructure.Common\Focus.Infrastructure.Common.csproj" />
+    <ProjectReference Include="..\Focus.Application.Common\Focus.Application.Common.csproj" />
   </ItemGroup>
 </Project>
