+++ src/Focus.Service.ReportProcessor/Application/Focus.Service.ReportProcessor.Application.csproj
@@ -5,6 +5,10 @@
     <ProjectReference Include="..\..\Focus.Application.Common\Focus.Application.Common.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="8.0.0" />
+  </ItemGroup>
+
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
